[FIX] 회원탈퇴, 카카오로그인 (#32)
* [FIX] 메소드 변환, 최신순 반영

* [FIX] 에러 수정

* [FIX] 순서 변경

* [FIX] 에러 수정

* [FEAT] 검색 api

* [FIX] 댓글 api 수정

* [FEAT] 프로필 사진 변경 api

* [FEAT] 유저탈퇴(삭제)api

* [FIX] 회원탈퇴 수정

* [FIX] 카카오 로그인 response 변경

+++ src/User/userController.js
@@ -72,6 +72,7 @@ const apple = new AppleAuth(appleAuthConfig, path.join(__dirname,'../../config/a
 
                     // 토큰 생성 성공
                     if (token) {
+                        
                         loginAgainResult.token = token;
                         return res.send(loginAgainResult);
                     }
@@ -104,11 +105,8 @@ const apple = new AppleAuth(appleAuthConfig, path.join(__dirname,'../../config/a
             
             // 토큰 생성 성공
             if (token) {
-                const tokenRes = {
-                    result,
-                    "token": token
-                };
-                return res.send(tokenRes);
+                result.token = token
+                return res.send(result);
             }
 
         }).catch(function (error) {
@@ -319,8 +317,18 @@ exports.deleteUser= async function(req, res) {
     const userIndex = token.userIndex;
 
     try {
-        const updateProfileUrlResult = await userService.deleteUser(userIndex);
-        return res.send(response(baseResponse.SUCCESS));
+        if(!userIndex) {
+            return res.json(errResponse(baseResponse.USER_USERID_EMPTY));
+        }
+
+        const selectUserResult = await userProvider.retrieveUserByuserIndex(userIndex);
+
+        if(!selectUserResult) {
+            return res.json(errResponse(baseResponse.USER_USERID_NOT_EXIST));
+        } else {
+            const updateProfileUrlResult = await userService.deleteUser(userIndex);
+            return res.send(response(baseResponse.SUCCESS));
+        }
     } catch (err) {
         console.log(err);
         logger.error(`사용자 삭제 중 Error`);
